Inserción de AUTOLOAD PSR-4
- Se insertó AUTOLOAD PSR-4 al código.
- Se renombraron los archivos de las clases.
- Se alineó el código para mayor comprensión.
- Se sustituyeron los require de las clases por los namespace.
